[SHLWAPI] Move all remaining kernelbase functions to kernelbase.c
This way the other files are almost completely identical to Wine-10 and once we export those functions from kernelbase, we can delete them from there.
